+++ R/autograph-package.R
@@ -1,0 +1,14 @@
+#' autograph: Automatic Plotting of Many Graphs
+#' @keywords internal
+#' @details
+#'   For more details, please consult the README and work through the tutorial.
+#'   The tutorial can be run locally using `run_tute("visualisation")`.
+#' @author James Hollway
+"_PACKAGE"
+
+# Global variables ####
+# defining global variables more centrally
+utils::globalVariables(c(".data", "obs", "valter","select","ego","zego","alter",
+                         "from","to","weight","unit","Step","Freq","Var1","n",
+                         "density","wave","period","name","value","sim","time",
+                         "q05","q95"))