+++ R/theme_palettes.R
@@ -125,3 +125,5 @@ match_color <- function(colors, pal) {
   matched
 }
 
+colorsafe_palette <- c("#d73027", "#4575b4", "#1B9E77","#D95F02","#7570B3",
+                       "#E7298A", "#66A61E","#E6AB02","#A6761D","#666666")