+++ tests/testthat/test-map_plot_manynet.R
@@ -1,5 +1,0 @@
-res <- manynet::node_deg(ison_dolphins)
-
-test_that("plotting degree distributions are the same", {
-  # expect_snapshot_file(ggsave("degdist.png", plot(res)), "degdist.png")
-})